Fixed warnings

+++ glm/gtc/ulp.inl
@@ -214,7 +214,7 @@ namespace glm
 #		if((GLM_LANG & GLM_LANG_CXX11_FLAG))
 			return std::nextafter(x, std::numeric_limits<double>::max());
 #		elif((GLM_COMPILER & GLM_COMPILER_VC) || ((GLM_COMPILER & GLM_COMPILER_INTEL) && (GLM_PLATFORM & GLM_PLATFORM_WINDOWS)))
-			return detail::nextafterf(x, std::numeric_limits<float>::max());
+			return detail::nextafterf(x, std::numeric_limits<double>::max());
 #		else
 			return nextafter(x, DBL_MAX);
 #		endif
@@ -243,7 +243,7 @@ namespace glm
 	GLM_FUNC_QUALIFIER double prev_float(double const & x)
 	{
 #		if((GLM_LANG & GLM_LANG_CXX11_FLAG))
-			return std::nextafter(x, std::numeric_limits<float>::min());
+			return std::nextafter(x, std::numeric_limits<double>::min());
 #		elif((GLM_COMPILER & GLM_COMPILER_VC) || ((GLM_COMPILER & GLM_COMPILER_INTEL) && (GLM_PLATFORM & GLM_PLATFORM_WINDOWS)))
 			return _nextafter(x, DBL_MIN);
 #		else
